conversational_agents_be/requirements.txt
livekit-agents[google]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
